[Yifan] Fix compiler warnings in pagerank.cpp and bench_common.h
- Cast RAND_MAX to float to avoid implicit int→float precision warning
- Cast hipEventDestroy return values to void to suppress nodiscard warnings

+++ gpu_perf_scripts/pagerank.cpp
@@ -55,7 +55,7 @@ int main(int argc, char** argv) {
     for (int i = 0; i < NUM_NODES; i++) {
         int nnz_row = 0;
         for (int j = 0; j < NUM_NODES; j++) {
-            float r = (float)rand() / RAND_MAX;
+            float r = (float)rand() / (float)RAND_MAX;
             if (r < SPARSITY) {
                 h_col.push_back((unsigned int)j);
                 nnz_row++;
